IOS delta time is fixed,Overflow in delta time bug is resolved,Bug is checking polyline visibility is resolved

+++ src/3d_shape/edge/base_edge/BaseEdge.h
@@ -49,7 +49,7 @@ private:
   Vec3DFloat triangleCenter = Vec3DFloat(0.0f,0.0f,0.0f);
   Vec3DFloat triangleCenterToPolygonCenterPoint = Vec3DFloat(0.0f,0.0f,0.0f);
   float dotProductValue = 0;
-  int normalVectorExtraDirectionFactor = 0;
+  float normalVectorExtraDirectionFactor = 0;
 };
 
 #endif
